engine/feat add engine skeleton

+++ src/ABStock.Shared/AgentState.cs
@@ -4,5 +4,5 @@ public record AgentState(
     string AgentName,
     AgentType AgentType,
     decimal Cash,
-    int Position
+    decimal Position
 );
